osd: fix subtitle/overlay update problems
This could in theory lead to missed updates if subtitles were switched
or external OSD overlays (via overlay-add) were updated. While the
change IDs of each of those were consistent, switching between two
separate OSD sources is not, and we have to explicitly trigger a change.

Regression since commit 2d51f7f3. The new code is actually better,
because we do exactly what is needed, and don't just mess with the
update ID for libass-based OSD.